requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
SQLAlchemy==1.4.23
alembic==1.7.3  # For database migrations (if using SQLAlchemy)
nltk==3.6.5
scikit-learn==0.24.2
gunicorn==21.2.0
python-dotenv==0.19.1
SpeechRecognition==3.8.1
pydantic~=1.10.13
passlib~=1.7.4
python-jose[cryptography]==3.3.0
